Fix dynamic version detection and git fetch-tags in CI workflow

+++ setup.py
@@ -44,27 +44,22 @@ Operating System :: MacOS
 ISRELEASED = True
 # assume a version set by conda, next update with git,
 # otherwise count on default
-VERSION = 'Unknown'
+VERSION = '0.0.1a1'
 
 
 class GitError(RuntimeError):
-    """Exception for git errors occuring in in git_version"""
+    """Exception for git errors occurring in git_version"""
     pass
 
 
 def git_version(srcdir=None):
-    """Return the git version, revision and cycle
-    Uses rev-parse to get the revision tag to get the version number from the
-    latest tag and detects (approximate) revision cycles
-    """
+    """Return the git version, revision and cycle"""
     def _minimal_ext_cmd(cmd, srcdir):
-        # construct minimal environment
         env = {}
         for k in ['SYSTEMROOT', 'PATH']:
             v = os.environ.get(k)
             if v is not None:
                 env[k] = v
-        # LANGUAGE is used on win32
         env['LANGUAGE'] = 'C'
         env['LANG'] = 'C'
         env['LC_ALL'] = 'C'
@@ -81,25 +76,39 @@ def git_version(srcdir=None):
                            % (proc.returncode, errmsg))
         return out
 
+    GIT_VERSION = VERSION
+    GIT_REVISION = 'Unknown'
+    GIT_CYCLE = 0
+
     try:
-        GIT_VERSION = VERSION
-        GIT_REVISION = 'Unknown'
-        GIT_CYCLE = 0
         out = _minimal_ext_cmd(['git', 'rev-parse', 'HEAD'], srcdir)
         GIT_REVISION = out.strip().decode('ascii')
+    except Exception:
+        pass
+
+    try:
         out = _minimal_ext_cmd(['git', 'tag'], srcdir)
-        GIT_VERSION = out.strip().decode('ascii').split('\n')[-1][0:]
+        tags = [t.strip() for t in out.strip().decode('ascii').split('\n') if t.strip()]
+        if tags:
+            GIT_VERSION = tags[-1]
+    except Exception:
+        pass
+
+    try:
         out = _minimal_ext_cmd(['git', 'describe', '--tags',
                                 '--long', '--always'], srcdir)
-        try:
-            # don't get a good description with shallow clones
-            GIT_CYCLE = out.strip().decode('ascii').split('-')[1]
-        except IndexError:
-            pass
-    except OSError:
+        desc = out.strip().decode('ascii').split('-')
+        if len(desc) >= 3:
+            GIT_CYCLE = desc[-2]
+            GIT_REVISION = desc[-1]
+        elif len(desc) == 1 and desc[0]:
+            GIT_CYCLE = 1
+            GIT_REVISION = desc[0]
+    except Exception:
         pass
 
     return GIT_VERSION, GIT_REVISION, GIT_CYCLE
+
 
 # BEFORE importing distutils, remove MANIFEST. distutils doesn't properly
 # update it when the contents of directories change.
@@ -108,11 +117,15 @@ if os.path.exists('MANIFEST'):
 
 
 def rewrite_setup_cfg(version, gitrevision, release):
+    cfg_in = 'setup.cfg.in'
+    if not os.path.exists(cfg_in):
+        return
     toreplace = dict(locals())
-    data = ''.join(open('setup.cfg.in', 'r').readlines()).split('@')
+    data = ''.join(open(cfg_in, 'r').readlines()).split('@')
     for k, v in toreplace.items():
-        idx = data.index(k)
-        data[idx] = v
+        if k in data:
+            idx = data.index(k)
+            data[idx] = str(v)
     cfg = open('setup.cfg', 'w')
     cfg.write(''.join(data))
     cfg.close()
@@ -122,49 +135,55 @@ def get_version_info(srcdir=None):
     global ISRELEASED
     GIT_CYCLE = 0
 
-    # Adding the git rev number needs to be done inside write_version_py(),
-    # otherwise the import messes up the build under Python 3.
+    git_dir = os.path.join(srcdir or '.', '.git')
+    setup_cfg_path = os.path.join(srcdir or '.', 'setup.cfg')
+
     if os.environ.get('CONDA_BUILD', False):
-        FULLVERSION = os.environ.get('PKG_VERSION', '???')
+        FULLVERSION = os.environ.get('PKG_VERSION', VERSION)
         GIT_REVISION = os.environ.get('GIT_DESCRIBE_HASH', '')
         ISRELEASED = True
         rewrite_setup_cfg(FULLVERSION, GIT_REVISION, 'yes')
-    elif os.path.exists('.git'):
+    elif os.path.exists(git_dir):
         FULLVERSION, GIT_REVISION, GIT_CYCLE = git_version(srcdir)
-        ISRELEASED = (GIT_CYCLE == 0 or GIT_CYCLE == '0')
+        ISRELEASED = (str(GIT_CYCLE) == '0')
+        if not FULLVERSION:
+            FULLVERSION = VERSION
         rewrite_setup_cfg(FULLVERSION, GIT_REVISION,
                           (ISRELEASED and 'yes') or 'no')
-    elif os.path.exists('setup.cfg'):
-        # valid distribution
+    elif os.path.exists(setup_cfg_path):
         setupcfg = configparser.ConfigParser(allow_no_value=True)
-        setupcfg.read('setup.cfg')
+        setupcfg.read(setup_cfg_path)
+        try:
+            FULLVERSION = setupcfg.get(section='metadata', option='version')
+        except Exception:
+            FULLVERSION = None
 
-        FULLVERSION = setupcfg.get(section='metadata', option='version')
+        if not FULLVERSION or FULLVERSION == "Unknown":
+            FULLVERSION = VERSION
 
-        if FULLVERSION is None:
-            FULLVERSION = "Unknown"
-
-        GIT_REVISION = setupcfg.get(section='metadata', option='gitrevision')
+        try:
+            GIT_REVISION = setupcfg.get(section='metadata', option='gitrevision')
+        except Exception:
+            GIT_REVISION = ""
 
         if GIT_REVISION is None:
             GIT_REVISION = ""
 
         return FULLVERSION, GIT_REVISION
     else:
-
-        # try to find a version number from the dir name
         dname = os.getcwd().split(os.sep)[-1]
-
         m = re.search(r'[0-9.]+', dname)
         if m:
             FULLVERSION = m.group()
             GIT_REVISION = ''
-
         else:
             FULLVERSION = VERSION
             GIT_REVISION = "Unknown"
 
-    if not ISRELEASED:
+    if not FULLVERSION:
+        FULLVERSION = VERSION
+
+    if not ISRELEASED and str(GIT_CYCLE) != '0':
         FULLVERSION += '.dev' + str(GIT_CYCLE)
 
     return FULLVERSION, GIT_REVISION
